Test to use case of generating new files
In this test the way the functionality is shown. The files that need to be
replicated in the generated code are marked with a special header, with the
symbol @ after the delimiter. Besides, it is expected that this special
anotation "implement" a function returning a list of objects with two properties:

* fileName: this value will be the fileName of the new file to create
* context: a value that will be injected to the next processor so it can be
used as context. In the files of the example the context will be the body of
the final files.

This header can be repeated any number of times, so complex filenames which
need more that one "value" of the specification json can be also generated.