Refactored DelimiterAwareEntity into Engine
Since we need to use functions like normalize both in processor and
file-generator, we need to place it in a common parent or in a helper file.
Since they both are engines, like the analyser, we have change the name of the
superclass into Engine and add a method providing this "helper" methods for
template generation (so far we only have normalize, but maybe there will be
more).