Logging and extra js functions
Now the derivation engine can get an optional parameter of a file where
a series of javascript functions can be defined to be shared between all the
files. This is, the functions can be used on the templates. This is useful to
validate or to generate code.

Also a new function is added, log(str) so a template can log whatever.